% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clin_subid.R
\name{clin_subid}
\alias{clin_subid}
\title{Reads the clinical data file}
\usage{
clin_subid(clinical_samplesheet, clin.samplesheet.case.id.column)
}
\arguments{
\item{clinical_samplesheet}{Clinical data matrix downloaded from TCGA or constructed with TCGAbiolinks (as shown in the example).}

\item{clin.samplesheet.case.id.column}{Column number containing the case IDs in the clinical file.}
}
\value{
A character vector.
}
\description{
Reads the clinical data file
}
