% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dstatus.R
\name{dstatus}
\alias{dstatus}
\title{Divides the disease status into binary group and creates a phenoytype matrix}
\usage{
dstatus(
  samplesheet,
  clinical_samplesheet,
  data.type1,
  samplesheet.data.type.column,
  samplesheet.case.id.column,
  clin.samplesheet.case.id.column,
  clin.samplesheet.stage.column,
  clin.samplesheet.age.column
)
}
\arguments{
\item{samplesheet}{Samplesheet matrix downloaded from TCGA or constructed with TCGAbiolinks (as shown in the example).}

\item{clinical_samplesheet}{Clinical data matrix downloaded from TCGA or constructed with TCGAbiolinks (as shown in the example).}

\item{data.type1}{Omic1 data type name (as in TCGA).}

\item{samplesheet.data.type.column}{Column number containing the data type in the samplesheet file.}

\item{samplesheet.case.id.column}{Column number containing the case IDs in the samplesheet file.}

\item{clin.samplesheet.case.id.column}{Column number containing the case IDs in the clinical file.}

\item{clin.samplesheet.stage.column}{Column number containing the tumor stage in the clinical file.}

\item{clin.samplesheet.age.column}{Column number containing the age values in the clinical file.}
}
\value{
A matrix.
}
\description{
Divides the disease status into binary group and creates a phenoytype matrix
}
