% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_pathwayGenes.R
\name{find_pathwayGenes}
\alias{find_pathwayGenes}
\title{Extracting pathway information for each gene from KEGG database}
\usage{
find_pathwayGenes()
}
\value{
A matrix with genes available from subject-specific files and the pathways they are involved in.
}
\description{
Extracting pathway information for each gene from KEGG database
}
\examples{
### Extracting pathway information for each gene from KEGG database

# library(AnnotationDbi)
# library(KEGGREST)
# library(org.Hs.eg.db)
# find_pathwayGenes()

# pathway.genes=find_pathwayGenes()

## DO NOT RUN
}
