% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iosearch_statistic.R
\name{iosearch_statistic}
\alias{iosearch_statistic}
\title{Calculates ioSearch test statistic}
\usage{
iosearch_statistic(full_data, K, ntopOmics2, ntopOmics1)
}
\arguments{
\item{full_data}{A list of omics1 and omics2 and phenotype data matrices}

\item{K}{Default is 1. Tuning parameter specific to elasticnet package.}

\item{ntopOmics2}{Number of top omics1 variables to be selected from each pathway (user-defined).}

\item{ntopOmics1}{Number of top omics1 variables to be selected from each pathway (user-defined).}
}
\value{
An array.
}
\description{
Calculates ioSearch test statistic
}
