% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omic1.R
\name{omic1}
\alias{omic1}
\title{Read the data and create a matrix of Omic1 (eg. protein) names by subjID}
\usage{
omic1(
  samplesheet,
  sampledata_folder,
  data.type1,
  sample.type,
  pattern,
  omic1.expr.column,
  samplesheet.sample.type.column,
  samplesheet.data.category.column,
  samplesheet.data.type.column
)
}
\arguments{
\item{samplesheet}{Samplesheet matrix downloaded from TCGA or constructed with TCGAbiolinks (as shown in the example).}

\item{sampledata_folder}{Omics data folder downloaded from TCGA or using TCGAbiolinks (as shown in the example).}

\item{data.type1}{Omic1 data type name (as in TCGA).}

\item{sample.type}{Sample type name (Tumor type such as "Primary Tumor", "Solid Tumor" etc.).}

\item{pattern}{String to extract omic1 data files.}

\item{omic1.expr.column}{Column number containing the omic1 expression values in the omic-specific datafile}

\item{samplesheet.sample.type.column}{Column number containing the sample type in the samplesheet file.}

\item{samplesheet.data.category.column}{Column number containing the data category in the samplesheet file.}

\item{samplesheet.data.type.column}{Column number containing the data type in the samplesheet file.}
}
\value{
A list.
}
\description{
Read the data and create a matrix of Omic1 (eg. protein) names by subjID
}
