% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omic1_module.R
\name{omic1_module}
\alias{omic1_module}
\title{Collect omic1 variables belonging to a set/pathway}
\usage{
omic1_module(x, mat, sets)
}
\arguments{
\item{x}{A numeric value.}

\item{mat}{Omic1 matrix.}

\item{sets}{Name of pathway.}
}
\value{
A matrix.
}
\description{
Collect omic1 variables belonging to a set/pathway
}
