% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Omic1_rmNA.R
\name{Omic1_rmNA}
\alias{Omic1_rmNA}
\title{Removes all missing data (such as NA, missing, no phenotype data) from omic1}
\usage{
Omic1_rmNA(
  samplesheet,
  sampledata_folder,
  clinical_samplesheet,
  data.type1,
  sample.type,
  pattern,
  omic1.expr.column,
  samplesheet.sample.type.column,
  samplesheet.data.category.column,
  samplesheet.case.id.column,
  samplesheet.data.type.column,
  clin.samplesheet.case.id.column,
  clin.samplesheet.stage.column,
  clin.samplesheet.age.column
)
}
\arguments{
\item{samplesheet}{Samplesheet matrix downloaded from TCGA or constructed with TCGAbiolinks (as shown in the example).}

\item{sampledata_folder}{Omics data folder downloaded from TCGA or using TCGAbiolinks (as shown in the example).}

\item{clinical_samplesheet}{Clinical data matrix downloaded from TCGA or constructed with TCGAbiolinks (as shown in the example).}

\item{data.type1}{Omic1 data type name (as in TCGA).}

\item{sample.type}{Sample type name (Tumor type such as "Primary Tumor", "Solid Tumor" etc.).}

\item{pattern}{String to extract omic1 data files.}

\item{omic1.expr.column}{Column number containing the omic1 expression values in the omic-specific datafile}

\item{samplesheet.sample.type.column}{Column number containing the sample type in the samplesheet file.}

\item{samplesheet.data.category.column}{A character vector with one element.}

\item{samplesheet.case.id.column}{Column number containing the case IDs in the samplesheet file.}

\item{samplesheet.data.type.column}{Column number containing the data type in the samplesheet file.}

\item{clin.samplesheet.case.id.column}{Column number containing the case IDs in the clinical file.}

\item{clin.samplesheet.stage.column}{Column number containing the tumor stage in the clinical file.}

\item{clin.samplesheet.age.column}{Column number containing the age values in the clinical file.}
}
\value{
A list.
}
\description{
Removes all missing data (such as NA, missing, no phenotype data) from omic1
}
