% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omic2.R
\name{omic2}
\alias{omic2}
\title{Create omic2 data matrix for both disease groups by collecting information from all subject-specific omic1 files.}
\usage{
omic2(
  samplesheet,
  sampledata_folder,
  pheno,
  subject.id.0,
  subject.id.1,
  data.type1,
  data.type2,
  sample.type,
  gene.type,
  pattern,
  gene.type.column,
  omic2.expr.column,
  samplesheet.data.category.column,
  samplesheet.sample.type.column,
  samplesheet.case.id.column,
  samplesheet.data.type.column
)
}
\arguments{
\item{samplesheet}{Samplesheet matrix downloaded from TCGA or constructed with TCGAbiolinks (as shown in the example).}

\item{sampledata_folder}{Omics data folder downloaded from TCGA or using TCGAbiolinks (as shown in the example).}

\item{pheno}{Phenotype file created from Omic1_rmNoOmic2().}

\item{subject.id.0}{Group1 subject IDs file created from Omic1_rmNoOmic2().}

\item{subject.id.1}{Group2 subject IDs file created from Omic1_rmNoOmic2().}

\item{data.type1}{Omic1 data type name (as in TCGA).}

\item{data.type2}{Omic2 data type name (as in TCGA).}

\item{sample.type}{Sample type name (Tumor type such as "Primary Tumor", "Solid Tumor" etc.).}

\item{gene.type}{Gene type of interest (eg. "protein_coding"). To be found in omics datafile. Could be an array as well.}

\item{pattern}{String to extract omic2 data files.}

\item{gene.type.column}{Column number containing the gene type in the omic-specific file.}

\item{omic2.expr.column}{Column number containing the omic2 expression values in the omic-specific datafile}

\item{samplesheet.data.category.column}{Column number containing the data category in the samplesheet file.}

\item{samplesheet.sample.type.column}{Column number containing the sample type in the samplesheet file.}

\item{samplesheet.case.id.column}{Column number containing the case IDs in the samplesheet file.}

\item{samplesheet.data.type.column}{Column number containing the data type in the samplesheet file.}
}
\value{
Individual matrices for omic2 expression data for each disease group.
}
\description{
Create omic2 data matrix for both disease groups by collecting information from all subject-specific omic1 files.
}
