% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omic2_module.R
\name{omic2_module}
\alias{omic2_module}
\title{Collect omic2 variables belonging to a set/pathway}
\usage{
omic2_module(x, mat, sets)
}
\arguments{
\item{x}{A numeric value.}

\item{mat}{Omic2 matrix.}

\item{sets}{Name of pathway.}
}
\value{
A matrix.
}
\description{
Collect omic2 variables belonging to a set/pathway
}
