% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathwayOmic1_expr.R
\name{pathwayOmic1_expr}
\alias{pathwayOmic1_expr}
\title{Collates pathway information with omic1 (that requires antibody map) data values for all subjects}
\usage{
pathwayOmic1_expr(
  pathway.genes,
  antibody_map,
  omic1.expr_Group1,
  omic1.expr_Group2
)
}
\arguments{
\item{pathway.genes}{Read the information obtained from find_pathwayGenes().}

\item{antibody_map}{Read the antibody map information directly downloaded from TCGA.}

\item{omic1.expr_Group1}{Read omic1 data matrix for all Group 1 subjects obtained from Omic1_rmNoOmic2().}

\item{omic1.expr_Group2}{Read omic1 data matrix for all Group 2 subjects obtained from Omic1_rmNoOmic2().}
}
\value{
A matrix with pathway information and omic1 expression values for all subjects.
}
\description{
Collates pathway information with omic1 (that requires antibody map) data values for all subjects
}
