% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subid.R
\name{subid}
\alias{subid}
\title{Finds the subject IDs}
\usage{
subid(
  samplesheet,
  data.type1,
  samplesheet.data.type.column,
  samplesheet.case.id.column
)
}
\arguments{
\item{samplesheet}{file downloaded from TCGA or prepared using TCGAbiolinks package}

\item{data.type1}{Select one data type to avoid subject ID repeats}

\item{samplesheet.data.type.column}{Indicate the data.type column number}

\item{samplesheet.case.id.column}{Indicate the case ID column number}
}
\value{
A character vector.
}
\description{
Finds the subject IDs
}
